issue #3 improvement: number of lines in methods reduced
